pyproject.toml
[tool.poetry]
name = "snowplow"
version = "0.0.1"
description = "Utilities for syncing data between snowflake and salesforce."
authors = ["rehket <aalbright@iamrehket.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
snowflake-connector-python = "^2.7.3"
httpx = "^0.22.0"
pydantic = "^1.9.0"
PyJWT = "^2.3.0"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
